+++ core/src/avm1/runtime.rs
@@ -6,7 +6,7 @@ use crate::avm1::object::TObject;
 use crate::avm1::property_map::PropertyMap;
 use crate::avm1::scope::Scope;
 use crate::avm1::{scope, Activation, ActivationIdentifier, Error, Object, Value};
-use crate::context::UpdateContext;
+use crate::context::{GcContext, UpdateContext};
 use crate::frame_lifecycle::FramePhase;
 use crate::prelude::*;
 use crate::string::AvmString;
@@ -73,8 +73,9 @@ pub struct Avm1<'gc> {
 }
 
 impl<'gc> Avm1<'gc> {
-    pub fn new(gc_context: MutationContext<'gc, '_>, player_version: u8) -> Self {
-        let (prototypes, globals, broadcaster_functions) = create_globals(gc_context);
+    pub fn new(context: &mut GcContext<'_, 'gc>, player_version: u8) -> Self {
+        let gc_context = context.gc_context;
+        let (prototypes, globals, broadcaster_functions) = create_globals(context);
 
         Self {
             player_version,
